requirements.txt
seaborn == 0.12.2
streamlit == 1.24.1
streamlit_option_menu == 0.3.6
pandas == 2.0.3
matplotlib == 3.7.2
plotly_express == 0.4.1
pillow == 10.0.0

